+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 - Improved docs for some MALClient functions
 
+## Changed
+- `get_auth_parts` now generates a random state
+- `get_auth_parts` now allows for a specified redirect uri
+- `auth` now takes a state and a redirect
+
 ## [v0.1.0]
 ### Added
 - Oauth2 authorization